% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongeElkan.R
\name{mongeElkan}
\alias{mongeElkan}
\title{Monge-Elkan Method for Approximate Text String Comparison}
\usage{
mongeElkan(string1, string2, method)
}
\arguments{
\item{string1}{character string to be matched}

\item{string2}{character string to be matched against}

\item{method}{Secondary string comparision method to be used}
}
\value{
Returns a comparision score in the range of 0 to 1 for the two strings,where 0 indicates no similarity between the two strings and 1 indicates a match between the 2 strings.
}
\description{
This algorithm extracts tokens (words sererated by whitespace) from the two input strings and then find the best matching pair of tokens in the set of tokens using a secondary similarity function.
}
\examples{
#Calculating mongeElkan score using jarowinkler as secondary comparision
mongeElkan("kaspersky antivirus","kasper antivirus","jw")
}
